Added some documentation, bugfixes and minor tweaks.
docs/ now contains the beginnings of documentation on the scripts.
plot_penultimate of custom/ was fixed so that it relies on the
spiketimes cell array to determine the number of units to plot instead
of the number of templates (which was significantly less than the actual
number of units). simple_snr was adjusted so that the threshold signal
to noise value is passed as an argument instead of being hard-coded to
the function.