Additions to documentation, bugfixes and parameter tweaks.
More scripts are now documented in docs/, but documentation is still far
from complete. A bugfix was added to get_sane to ensure that it
functions correctly when units are discarded becuase they lack the
sufficient wave count. Originally this could cause a problem where an
array contains uninitialized values. Finally, get_sane's coefficient of
variation parameter has been changed from 0.4 to 0.25 in an attempt to
be more strict in removing artifacts.