Further additions to documentation and tuning.
get_sane and pca_merge have further refinements with regard to the
percentage difference tolerance in the max and min voltage of units
(pca_merge) units as well as the coefficient of
variation tolerance (get_sane). "manual.pdf" now contains a basic
summary for all of the functions and script and examples of basic usage
for the most relevant subroutines.