Added parameters, basic logic to choose sorting routine.
spikesort_muxi now has two parameters, do_get_sane and do_pca_merge,
which cause get_final_units to use get_sane and pca_merge when asserted,
respectively. These options are asserted when set to 'y' and deasserted
when set to 'n'. Not that combinations of the two parameters may be
used--they do not have to be both asserted or deasserted at the same
time.